week4_library_demo/Backend/transaction_service/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
mysql-connector-python==9.4.0
requests==2.32.3